Removed merge check from PG, formatting
