+++ yb-voyager/cmd/exportSchema.go
@@ -20,11 +20,9 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/callhome"
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/utils"
-	"golang.org/x/exp/slices"
 
 	"github.com/spf13/cobra"
 )
@@ -81,15 +79,6 @@ func exportSchema() {
 
 	CreateMigrationProjectIfNotExists(source.DBType, exportDir)
 	source.DB().ExportSchema(exportDir)
-	schemaList := utils.GetSchemaObjectList(source.DBType)
-	for _, objectType := range schemaList {
-		if !slices.Contains([]string{"VIEW", "MVIEW"}, objectType) {
-			continue
-		}
-		if err = processImportDirectives(filepath.Join(exportDir, "schema", strings.ToLower(objectType)+"s", strings.ToLower(objectType)+".sql")); err != nil {
-			utils.ErrExit(err.Error())
-		}
-	}
 	utils.PrintAndLog("\nExported schema files created under directory: %s\n", filepath.Join(exportDir, "schema"))
 
 	payload := callhome.GetPayload(exportDir)
